+++ LD45/Assets/Scripts/FlyingEnemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4413b665796753a4c98c8d8afcbcb911
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
