+++ LD45/Assets/Scripts/FlyingEnemy/FlyingEnemy.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 13a65b02d7cbe82458d07cd3d35aef59
+guid: be4781d9352cf444a9bd86cec1a3089d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
