+++ LD45/Assets/Scripts/FlyingEnemy/FlyingEnemyState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b4b2f39831823e4aa37537032b84c7e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
