+++ LD45/Assets/Scripts/FlyingEnemy/FlyingNavAvent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f5f0eef31abe624469ad772a1bdd8ec0
+guid: 345a8b329e701a24d8a6b39b17d2a17a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
